とりあえずtrapbin.cを分離
