+++ trapsfc.c
@@ -1,2 +1,8 @@
+#include <stdio.h>
+#include <string.h>
+#include <math.h>
+#include <time.h>
+#include "mymalloc.h"
+#include "visual.h"
+#include "gribscan.h"
 #include "grib2png.h"
-
